Mescaldos join

+++ cardinal_app/src/main/java/cu/phibrain/cardinal/app/helpers/LatLongUtils.java
@@ -21,7 +21,7 @@ import eu.geopaparazzi.map.GPGeoPoint;
 public class LatLongUtils {
     public static final double MAX_DISTANCE = 50.0f;
     public static final double LINE_AND_POLYGON_VIEW_ZOOM = 15;
-    public static final int RADIUS_JOIN_MO = 100;
+    public static final float RADIUS_JOIN_MO = 100.0f;
     public static double distance(MapObject mo1, MapObject mo2) {
         try {
             GPGeoPoint p1 = mo1.getCoord().get(mo1.getCoord().size() - 1);
